requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.6
alembic>=1.11.1
fastapi>=0.97.0
uvicorn>=0.22.0
pytest>=7.3.2
fastapi-mail>=1.2.8
pydantic>=1.10.9
py-automapper>=1.2.3
jinja2>=3.1.2
anyio>=3.7.0
httpx>=0.24.1
passlib[bcrypt]>=1.7.4
dramatiq[rabbitmq, watch]>=1.14.2
typeguard>=4.0.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0
sqlalchemy>=2.0.18
schedule>=1.2.0